review: simplify and fix UC-002 implementation
- Skip sending empty message in ThreadCreator when no optional fields provided
- Move slash command handler logic from main.go to gateway/discord/new_order_handler.go
- Add nil-safe caller ID extraction (guards against nil Member in DM context)
