+++ main.go
@@ -15,7 +15,6 @@ import (
 	"github.com/jomei/notionapi"
 
 	"github.com/xgnid-tw/gx5/config"
-	"github.com/xgnid-tw/gx5/domain"
 	discordgw "github.com/xgnid-tw/gx5/gateway/discord"
 	notiongw "github.com/xgnid-tw/gx5/gateway/notion"
 	"github.com/xgnid-tw/gx5/usecase"
@@ -99,9 +98,7 @@ func main() {
 	cmdHandler := discordgw.NewCommandHandler(dc, cfg.DiscordAppID)
 	defer cmdHandler.UnregisterAll()
 
-	err = cmdHandler.RegisterCommand(newOrderCommand(), func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-		handleNewOrder(s, i, createOrderUC)
-	})
+	err = cmdHandler.RegisterCommand(discordgw.NewOrderCommand(), discordgw.HandleNewOrder(createOrderUC))
 	if err != nil {
 		log.Fatalf("error registering newOrder command: %s", err)
 	}
@@ -116,86 +113,4 @@ func main() {
 	<-sc
 
 	_ = s.Shutdown()
-}
-
-func newOrderCommand() *discordgo.ApplicationCommand {
-	return &discordgo.ApplicationCommand{
-		Name:        "neworder",
-		Description: "Create a new group purchase order",
-		Options: []*discordgo.ApplicationCommandOption{
-			{
-				Type:        discordgo.ApplicationCommandOptionString,
-				Name:        "ordertitle",
-				Description: "Name of the order",
-				Required:    true,
-			},
-			{
-				Type:        discordgo.ApplicationCommandOptionString,
-				Name:        "deadline",
-				Description: "Order deadline (YYYY-MM-DD)",
-				Required:    false,
-			},
-			{
-				Type:        discordgo.ApplicationCommandOptionString,
-				Name:        "shopurl",
-				Description: "Shop URL",
-				Required:    false,
-			},
-			{
-				Type:        discordgo.ApplicationCommandOptionString,
-				Name:        "tags",
-				Description: "Tag (315pro, 学マス, 283pro, 346pro, 765pro)",
-				Required:    false,
-			},
-		},
-	}
-}
-
-func handleNewOrder(s *discordgo.Session, i *discordgo.InteractionCreate, uc *usecase.CreateOrder) {
-	opts := i.ApplicationCommandData().Options
-	optMap := make(map[string]*discordgo.ApplicationCommandInteractionDataOption, len(opts))
-
-	for _, opt := range opts {
-		optMap[opt.Name] = opt
-	}
-
-	order := domain.Order{}
-
-	if v, ok := optMap["ordertitle"]; ok {
-		order.ThreadName = v.StringValue()
-	}
-
-	if v, ok := optMap["deadline"]; ok {
-		order.Deadline = v.StringValue()
-	}
-
-	if v, ok := optMap["shopurl"]; ok {
-		order.ShopURL = v.StringValue()
-	}
-
-	if v, ok := optMap["tags"]; ok {
-		order.Tag = domain.Tag(v.StringValue())
-	}
-
-	callerID := i.Member.User.ID
-
-	err := uc.Execute(context.Background(), callerID, i.ChannelID, order)
-	if err != nil {
-		respondToInteraction(s, i, "Error: "+err.Error())
-		return
-	}
-
-	respondToInteraction(s, i, "Order created: "+order.ThreadName)
-}
-
-func respondToInteraction(s *discordgo.Session, i *discordgo.InteractionCreate, message string) {
-	err := s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
-		Type: discordgo.InteractionResponseChannelMessageWithSource,
-		Data: &discordgo.InteractionResponseData{
-			Content: message,
-		},
-	})
-	if err != nil {
-		log.Printf("error responding to interaction: %s", err)
-	}
 }
